+++ coronataiou/app.py
@@ -3,14 +3,11 @@
 from flask import Flask, render_template, request, flash, url_for,  jsonify
 from flask_sqlalchemy import SQLAlchemy
 from flask_bootstrap import Bootstrap
+from datetime import datetime, timedelta, date
 
-from flask_wtf import FlaskForm
-from wtforms import SubmitField, SelectField, RadioField, HiddenField, StringField, IntegerField, FloatField
-from wtforms.validators import InputRequired, Length, Regexp, NumberRange
-from datetime import date
-
+from coronataiou.form import AddRecord, DeleteForm
 from coronataiou.models import db, ma, RecordData, RecordSchema
-from datetime import datetime, timedelta
+
 
 
 app = Flask(__name__)
@@ -28,31 +25,7 @@ Bootstrap(app)
 # form for add_record and edit_or_delete
 # each field includes validation requirements and messages
 
-class AddRecord(FlaskForm):
-    id_field = HiddenField()
-    name = StringField('Name', [InputRequired(),
-                                Regexp(r'^[A-Za-z\s\-\']+$', message="Invalid name"),
-                                Length(min=3, max=25, message="Invalid name")
-                                ])
-    temperature = FloatField('temperature ℃', [InputRequired(),
-                                               NumberRange(min=1.00, max=99.99, message="Invalid range")
-                                               ])
-    fatigue = SelectField('Do you feel fatigue ', [InputRequired()],
-                          choices=[('', ''), ('Yes', 'Yes'),('No','No') ])
 
-    sore_throat = SelectField('Does your feel sore hurt? ', [InputRequired()],
-                              choices=[('', ''), ('Yes', 'Yes'),('No','No') ])
-    other_pain = SelectField('Do you have any other pain? ', [InputRequired()],
-                             choices=[('', ''), ('Yes', 'Yes'),('No','No') ])
-
-    update = HiddenField()
-    submit = SubmitField('Add/Update Record')
-
-
-class DeleteForm(FlaskForm):
-    id_filed = HiddenField()
-    purpose = HiddenField()
-    submit = SubmitField('Delete this data')
 
 
 def stringdate():
@@ -82,7 +55,8 @@ def add_record():
         other_pain = request.form['other_pain']
 
         updated = stringdate()
-        record_datas = record_data(name,temperature,fatigue,sore_throat,other_pain,updated)
+
+        record_datas = RecordData(name, temperature, fatigue, sore_throat, other_pain, updated)
 
         db.session.add(record_datas)
         db.session.commit()
@@ -102,73 +76,73 @@ def add_record():
         return render_template('add_record_temperature.html', form1=form1)
 
 # select a record to edit or delete
-@app.route('/select_record/<letters>')
-def select_record(letters):
-    # alphabetical lists by name, chunked by letters between _ and _
-    # .between() evaluates first letter of a string
-    a, b = list(letters)
-    record_datas = record_data.query.filter(record_data.name.between(a, b)).order_by(record_data.name).all()
-    return render_template('select_record_temperature.html', record_data=record_datas)
-
-# edit or delete - come here from form in /select_record
-@app.route('/edit_or_delete', methods=['POST'])
-def edit_or_delete():
-    id = request.form['id']
-    choice = request.form['choice']
-    record_datas = record_data.query.filter(record_data.id == id).first()
-    # two forms in this template
-    form1 = AddRecord()
-    form2 = DeleteForm()
-    return render_template('edit_or_delete_temperature.html', record_data=record_datas, form1=form1, form2=form2, choice=choice)
-
-
-@app.route('/delete_result', methods=['POST'])
-def delete_result():
-    id = request.form['id_field']
-    purpose = request.form['purpose']
-    record_datas = record_data.query.filter(record_data.id == id).first()
-    if purpose == 'delete':
-        db.session.delete(record_datas)
-        db.session.commit()
-        message = f"The data {record_datas.name} has been deleted from the database."
-        return render_template('result.html', message=message)
-    else:
-        # this calls an error handler
-        abort(405)
+# @app.route('/select_record/<letters>')
+# def select_record(letters):
+#     # alphabetical lists by name, chunked by letters between _ and _
+#     # .between() evaluates first letter of a string
+#     a, b = list(letters)
+#     record_datas = record_data.query.filter(record_data.name.between(a, b)).order_by(record_data.name).all()
+#     return render_template('select_record_temperature.html', record_data=record_datas)
+#
+# # edit or delete - come here from form in /select_record
+# @app.route('/edit_or_delete', methods=['POST'])
+# def edit_or_delete():
+#     id = request.form['id']
+#     choice = request.form['choice']
+#     record_datas = record_data.query.filter(record_data.id == id).first()
+#     # two forms in this template
+#     form1 = AddRecord()
+#     form2 = DeleteForm()
+#     return render_template('edit_or_delete_temperature.html', record_data=record_datas, form1=form1, form2=form2, choice=choice)
+#
+#
+# @app.route('/delete_result', methods=['POST'])
+# def delete_result():
+#     id = request.form['id_field']
+#     purpose = request.form['purpose']
+#     record_datas = record_data.query.filter(record_data.id == id).first()
+#     if purpose == 'delete':
+#         db.session.delete(record_datas)
+#         db.session.commit()
+#         message = f"The data {record_datas.name} has been deleted from the database."
+#         return render_template('result.html', message=message)
+#     else:
+#         # this calls an error handler
+#         abort(405)
 
 # result of edit - this function updates the record
-@app.route('/edit_result', methods=['POST'])
-def edit_result():
-    id = request.form['id_field']
-    # call up the record from the database
-    record_datas = record_data.query.filter(record_data.id == id).first()
-    # update all values
-    record_datas.name = request.form['name']
-    record_datas.temperature = request.form['temperature']
-    record_datas.fatigue = request.form['fatigue']
-    record_datas.sore_throat = request.form['sore_throat']
-    record_datas.other_pain = request.form['other_pain']
-    # get today's date from function, above all the routes
-    record_datas.updated = stringdate()
-
-    form1 = AddRecord()
-    if form1.validate_on_submit():
-        # update database record
-        db.session.commit()
-        # create a message to send to the template
-        message = f"The data for {record_datas.name} has been updated."
-        return render_template('result.html', message=message)
-    else:
-        # show validaton errors
-        record_datas.id = id
-        # see https://pythonprogramming.net/flash-flask-tutorial/
-        for field, errors in form1.errors.items():
-            for error in errors:
-                flash("Error in {}: {}".format(
-                    getattr(form1, field).label.text,
-                    error
-                ), 'error')
-        return render_template('edit_or_delete.html', form1=form1, record_data=record_datas, choice='edit')
+# @app.route('/edit_result', methods=['POST'])
+# def edit_result():
+#     id = request.form['id_field']
+#     # call up the record from the database
+#     record_datas = record_data.query.filter(record_data.id == id).first()
+#     # update all values
+#     record_datas.name = request.form['name']
+#     record_datas.temperature = request.form['temperature']
+#     record_datas.fatigue = request.form['fatigue']
+#     record_datas.sore_throat = request.form['sore_throat']
+#     record_datas.other_pain = request.form['other_pain']
+#     # get today's date from function, above all the routes
+#     record_datas.updated = stringdate()
+#
+#     form1 = AddRecord()
+#     if form1.validate_on_submit():
+#         # update database record
+#         db.session.commit()
+#         # create a message to send to the template
+#         message = f"The data for {record_datas.name} has been updated."
+#         return render_template('result.html', message=message)
+#     else:
+#         # show validaton errors
+#         record_datas.id = id
+#         # see https://pythonprogramming.net/flash-flask-tutorial/
+#         for field, errors in form1.errors.items():
+#             for error in errors:
+#                 flash("Error in {}: {}".format(
+#                     getattr(form1, field).label.text,
+#                     error
+#                 ), 'error')
+#         return render_template('edit_or_delete.html', form1=form1, record_data=record_datas, choice='edit')
 
 @app.errorhandler(404)
 def page_not_found(e):
